Updating _xhrTransport to support File upload
Made FileUploader._xhrTransport a bit more intelligent about the type of upload to be performed.  If form data was provided for the item, then a FormData upload is made, else a simple File upload is performed instead.  This gives the browser a chance to correctly set the Expect: 100-continue header for large files (a big deal for storage servers that may need a chance to do special processing for larger files).