Additional FormData fix
Additional FormData for files was being looped twice causing form data
to be sent as individual characters in an array. This fixes that issue
and now form data on the upload is an Object instead of an array stated
in the Docs. You would have to ix the documentation as well stating
that formData is an Object containing key/value pairs to be uploaded as
part of the request body.